backend/laweb/pyproject.toml
[tool.poetry]
name = "laweb backend"
version = "0.1.0"
description = "Backend for Laweb module"
authors = ["Flict-dev <zhdanom.maksi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aioredis = "^2.0.1"
fastapi = "^0.90.1"
pytest = "^7.2.1"
python-decouple = "^3.7"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
